Pos ya le entendi mas :v

+++ Facturacion/database/migrations/2018_06_19_025013_create_impuestos_table.php
@@ -16,7 +16,7 @@ class CreateImpuestosTable extends Migration
         Schema::create('impuestos', function (Blueprint $table) {
             $table->increments('id');
             $table->string('clave', 200);
-            $table->decimal('impuesto');
+            $table->float('impuesto');
             $table->timestamps();
         });
     }
